refactor: Improve structure and security
- Isolate main process logic into separate modules (docker, windows, tray, updater)
- Centralize configuration into a dedicated config.js file
- Fix Content Security Policy by moving inline scripts and styles to external files
- Secure main window by enabling contextIsolation and disabling nodeIntegration
