+++ src/main/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  APP_URL: "http://localhost:3000",
+  SERVICE_NAME: "web",
+  WAIT_TIMEOUT: 60_000, // 60 seconds
+};
