+++ src/main/windows.js
@@ -1,0 +1,60 @@
+const { BrowserWindow } = require("electron");
+const path = require("path");
+const { APP_URL } = require("./config");
+
+let loadingWindow;
+let mainWindow;
+
+function createLoadingWindow() {
+  loadingWindow = new BrowserWindow({
+    width: 420,
+    height: 280,
+    frame: false,
+    resizable: false,
+    webPreferences: {
+      preload: path.join(__dirname, "../preload/preload.js"),
+      nodeIntegration: false,
+      contextIsolation: true,
+    },
+  });
+  loadingWindow.loadFile(path.join(__dirname, "../renderer/loading.html"));
+  return loadingWindow;
+}
+
+function showError(message) {
+  if (!loadingWindow || loadingWindow.isDestroyed()) return;
+  loadingWindow.webContents.send("error", message);
+  loadingWindow.loadFile(path.join(__dirname, "../renderer/error.html"));
+}
+
+function createMainWindow() {
+  mainWindow = new BrowserWindow({
+    width: 1200,
+    height: 800,
+    webPreferences: {
+      preload: path.join(__dirname, "../preload/preload.js"),
+      nodeIntegration: false,
+      contextIsolation: true,
+    },
+  });
+
+  mainWindow.loadURL(APP_URL);
+
+  mainWindow.on("close", (e) => {
+    if (!quitting) {
+      e.preventDefault();
+      mainWindow.hide();
+    }
+  });
+
+  loadingWindow?.close();
+  return mainWindow;
+}
+
+module.exports = {
+  createLoadingWindow,
+  showError,
+  createMainWindow,
+  getMainWindow: () => mainWindow,
+  getLoadingWindow: () => loadingWindow,
+};
