+++ src/renderer/error.html
@@ -2,28 +2,13 @@
 <html>
 <head>
     <meta charset="UTF-8">
+    <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'; style-src 'self'">
     <title>Panacea Launcher</title>
-    <style>
-        body {
-            font-family: sans-serif;
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-            color: red;
-        }
-    </style>
+    <link rel="stylesheet" href="css/error.css">
 </head>
 <body>
     <h1>Error</h1>
     <p id="error"></p>
-    <script>
-        const error = document.getElementById('error');
-        window.electronAPI.onError((event, message) => {
-            error.innerText = message;
-        });
-    </script>
+    <script src="js/error.js"></script>
 </body>
 </html>
