+++ src/renderer/js/error.js
@@ -1,0 +1,4 @@
+const error = document.getElementById('error');
+window.electronAPI.onError((event, message) => {
+    error.innerText = message;
+});
