+++ src/renderer/js/loading.js
@@ -1,0 +1,4 @@
+const status = document.getElementById('status');
+window.electronAPI.onStatus((event, message) => {
+    status.innerText = message;
+});
